+++ src/models/light_cone/LightConeSuperimposition.ts
@@ -52,7 +52,7 @@ class LightConeSuperimposition {
                 return StatProperty.ALL_DAMAGE_TYPES.map(damageType => new StatPropertyValue(damageType, prop.getAsNumber("Value", "Value"), this.client));
             }
             return [new StatPropertyValue(prop.getAsString("PropertyType") as StatPropertyType, prop.getAsNumber("Value", "Value"), this.client)];
-        }).reduce((a, b) => [...a, ...b], []);
+        }).flat();
     }
 }