+++ src/models/relic/RelicSet.ts
@@ -1,9 +1,11 @@
-import { JsonObject, JsonReader } from "config_file.js";
+import { JsonObject, JsonReader, separateByValue } from "config_file.js";
 import StarRail from "../../client/StarRail";
 import AssetsNotFoundError from "../../errors/AssetsNotFoundError";
 import TextAssets from "../assets/TextAssets";
 import ImageAssets from "../assets/ImageAssets";
 import RelicSetBonus from "./RelicSetBonus";
+import Relic from "./Relic";
+import RelicData from "./RelicData";
 
 /**
  * @en RelicSet
@@ -45,6 +47,25 @@ class RelicSet {
         this.figureIcon = new ImageAssets(json.getAsString("SetIconFigurePath"), this.client);
 
         this.setBonus = json.get("SetSkillList").mapArray((_, needCount) => new RelicSetBonus(this.id, needCount.getAsNumber(), this.client));
+    }
+
+    /**
+    * @param relics
+    */
+    static getActiveSetBonus(relics: (Relic | RelicData | RelicSet)[]): { set: RelicSet, count: number, activeBonus: RelicSetBonus[] }[] {
+        const relicSets = relics.map(a => (a instanceof RelicSet) ? a : (a instanceof RelicData) ? a.set : a.relicData.set);
+
+        const separated = separateByValue(relicSets, a => a.id.toString());
+
+        const relicSetCounts = Object.values(separated).map(array => { return { count: array.length, set: array[0] }; });
+
+        return relicSetCounts.map(obj => {
+            return {
+                set: obj.set,
+                count: obj.count,
+                activeBonus: obj.set.setBonus.filter(bonus => bonus.needCount <= obj.count),
+            };
+        }).sort((a, b) => b.count - a.count);
     }
 }