+++ src/models/relic/RelicSetBonus.ts
@@ -1,6 +1,9 @@
-import { JsonObject } from "config_file.js";
+import { JsonObject, JsonReader } from "config_file.js";
 import StarRail from "../../client/StarRail";
 import AssetsNotFoundError from "../../errors/AssetsNotFoundError";
+import TextAssets from "../assets/TextAssets";
+import { getStableHash } from "../../utils/hash_utils";
+import { StatPropertyType, StatPropertyValue } from "../StatProperty";
 
 /**
  * @en RelicSetBonus
@@ -12,6 +15,11 @@ class RelicSetBonus {
     readonly needCount: number;
     /**  */
     readonly client: StarRail;
+
+    /**  */
+    readonly description: TextAssets;
+    /**  */
+    readonly stats: StatPropertyValue[];
 
     readonly _data: JsonObject;
 
@@ -29,7 +37,17 @@ class RelicSetBonus {
         if (!_data) throw new AssetsNotFoundError("RelicSetBonus", `${this.setId}-${this.needCount}`);
         this._data = _data as JsonObject;
 
-        // const json = new JsonReader(this._data);
+        const json = new JsonReader(this._data);
+
+        this.description = new TextAssets(getStableHash(json.getAsString("SkillDesc")), this.client);
+
+        this.stats = json.get("PropertyList").mapArray((_, stat) => new StatPropertyValue(
+            stat.getAsString(this.client.cachedAssetsManager.getObjectKeysManager().relicSetBonusStatPropertyTypeKey) as StatPropertyType,
+            stat.getAsNumber(this.client.cachedAssetsManager.getObjectKeysManager().relicSetBonusStatPropertyValueKey, "Value"),
+            this.client,
+        ));
+
+
     }
 }